+++ examples/ch03/hello.c
@@ -1,0 +1,7 @@
+#include <stdio.h>
+
+int main(void)
+{
+    printf("안녕, 세상!\n");
+    return 0;
+}